requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
fastapi==0.128.4
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageHash==4.3.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas==13.0.0.19
nvidia-cuda-cupti==13.0.48
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-runtime==13.0.48
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.3.29
nvidia-cusparse==12.6.2.49
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytesseract==0.3.13
pytest==9.0.2
python-multipart==0.0.22
PyWavelets==1.9.0
scipy==1.17.0
starlette==0.52.1
sympy==1.14.0
torch==2.9.1+cu130
torchvision==0.24.1+cu130
tqdm==4.67.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uv==0.9.25
uvicorn==0.40.0
